refactor(collector): move mock implementation from impl/ to mock/
MockModule, MockParams and the mock implementation lived in impl/ which
is not allowed to import go.uber.org/fx or fxutil per the V2 component
convention. Move them to mock/mock.go which is the designated location
for mock definitions that need fx wiring.

Callers referencing collectorimpl.MockParams are updated to use
collectormock.MockParams from the mock package.
